+++ src/main/kotlin/com/merkost/drawablepreview/DrawablePreviewIconProvider.kt
@@ -1,0 +1,10 @@
+package com.merkost.drawablepreview
+
+import com.intellij.ide.IconProvider
+import com.intellij.psi.PsiElement
+import com.merkost.drawablepreview.factories.IconPreviewFactory
+
+class DrawablePreviewIconProvider : IconProvider() {
+
+    override fun getIcon(element: PsiElement, flags: Int) = IconPreviewFactory.createIcon(element)
+}
