+++ src/main/kotlin/com/merkost/drawablepreview/Extensions.kt
@@ -1,0 +1,7 @@
+package com.merkost.drawablepreview
+
+fun String.getDigits() = try {
+    this.replace("\\D+", "").toInt()
+} catch (exception: Exception) {
+    null
+}
