+++ src/main/kotlin/com/merkost/drawablepreview/drawables/dom/Drawable.kt
@@ -1,0 +1,13 @@
+package com.merkost.drawablepreview.drawables.dom
+
+import org.w3c.dom.Element
+import java.awt.image.BufferedImage
+
+abstract class Drawable {
+
+    open fun inflate(element: Element) {
+    }
+
+    open fun draw(outputImage: BufferedImage) {
+    }
+}
