--- src/main/kotlin/com/mistamek/drawablepreview/Extensions.kt
@@ -1,7 +1,0 @@
-package com.mistamek.drawablepreview
-
-fun String.getDigits() = try {
-    this.replace("\\D+", "").toInt()
-} catch (exception: Exception) {
-    null
-}
