--- src/main/kotlin/com/mistamek/drawablepreview/factories/RetinaImageIcon.kt
@@ -1,15 +1,0 @@
-package com.mistamek.drawablepreview.factories
-
-import com.intellij.util.ui.UIUtil
-import java.awt.Component
-import java.awt.Graphics
-import java.awt.Image
-import javax.swing.ImageIcon
-
-class RetinaImageIcon(image: Image) : ImageIcon(image, "") {
-
-    @Synchronized
-    override fun paintIcon(c: Component, g: Graphics, x: Int, y: Int) {
-        UIUtil.drawImage(g, this.image, x, y, null)
-    }
-}
